requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.5
pytest-base-url==2.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
SQLAlchemy==2.0.36
sqlparse==0.5.2
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
Werkzeug==3.1.3
